Improved messages and clean network class
