+++ autoPyTorch/pipeline/components/setup/lr_scheduler/CosineAnnealingLR.py
@@ -45,7 +45,7 @@ class CosineAnnealingLR(BaseLRComponent):
         """
 
         # Make sure there is an optimizer
-        self.check_requirements(X)
+        self.check_requirements(X, y)
 
         self.scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(
             optimizer=X['optimizer'],
