+++ autoPyTorch/pipeline/components/setup/lr_scheduler/ExponentialLR.py
@@ -46,7 +46,7 @@ class ExponentialLR(BaseLRComponent):
         """
 
         # Make sure there is an optimizer
-        self.check_requirements(X)
+        self.check_requirements(X, y)
 
         self.scheduler = torch.optim.lr_scheduler.ExponentialLR(
             optimizer=X['optimizer'],
