+++ autoPyTorch/pipeline/components/setup/lr_scheduler/StepLR.py
@@ -51,7 +51,7 @@ class StepLR(BaseLRComponent):
         """
 
         # Make sure there is an optimizer
-        self.check_requirements(X)
+        self.check_requirements(X, y)
 
         self.scheduler = torch.optim.lr_scheduler.StepLR(
             optimizer=X['optimizer'],
