+++ autoPyTorch/pipeline/components/setup/network/base_network.py
@@ -1,5 +1,8 @@
 import numbers
-from typing import Any, Dict
+from abc import abstractmethod
+from typing import Any, Dict, Optional
+
+import numpy as np
 
 import torch
 
@@ -10,8 +13,57 @@ class BaseNetworkComponent(autoPyTorchSetupComponent):
     """Provide an abstract interface for networks
     in Auto-Pytorch"""
 
-    def __init__(self) -> None:
+    def __init__(
+        self,
+        num_layers: int,
+        intermediate_activation: str,
+        use_dropout: bool,
+        final_activation: Optional[str],
+        random_state: Optional[np.random.RandomState] = None,
+    ) -> None:
         self.network = None
+        self.num_layers = num_layers
+        self.intermediate_activation = intermediate_activation
+        self.use_dropout = use_dropout
+        self.random_state = random_state
+
+    def fit(self, X: Dict[str, Any], y: Any = None) -> autoPyTorchSetupComponent:
+        """
+        Fits a component by using an input dictionary with pre-requisites
+
+        Args:
+            X (X: Dict[str, Any]): Dependencies needed by current component to perform fit
+            y (Any): not used. To comply with sklearn API
+
+        Returns:
+            A instance of self
+        """
+        # Make sure that input dictionary X has the required
+        # information to fit this stage
+        self.check_requirements(X, y)
+
+        in_features = X['num_features']
+        out_features = X['num_classes']
+
+        self.network = self.build_network(in_features, out_features)
+
+        return self
+
+    @abstractmethod
+    def build_network(self, in_feature: int, out_features: int) -> torch.nn.Module:
+        """This method returns a pytorch network, that is dynamically built
+        using:
+
+            common network arguments from the base class:
+                * num_layer
+                * intermediate_activation
+                * use_dropout
+                * final_activation
+
+            a self.config that is network specific, and contains the additional
+            configuration hyperparameters to build a domain specific network
+        """
+        raise NotImplementedError()
 
     def transform(self, X: Dict[str, Any]) -> Dict[str, Any]:
         """The transform function calls the transform function of the
@@ -31,17 +83,17 @@ class BaseNetworkComponent(autoPyTorchSetupComponent):
         Returns:
             model : the underlying network object
         """
-        assert self.network is not None, "No network was fit"
+        assert self.network is not None, "No network was initialized"
         return self.network
 
-    def check_requirements(self, X: Dict[str, Any]) -> None:
+    def check_requirements(self, X: Dict[str, Any], y: Any = None) -> None:
         """ This common utility makes sure that the input dictionary X,
         used to fit a given component class, contains the minimum information
         to fit the given component, and it's parents
         """
 
         # Honor the parent requirements
-        super().check_requirements(X)
+        super().check_requirements(X, y)
 
         # For the Network, we need the number of input features,
         # to build the first network layer
@@ -92,6 +144,11 @@ class BaseNetworkComponent(autoPyTorchSetupComponent):
             'elu': torch.nn.ELU,
             'prelu': torch.nn.PReLU,
         }
+
+    def get_network_weights(self) -> torch.nn.parameter.Parameter:
+        """Returns the weights of the network"""
+        assert self.network is not None, "No network was initialized"
+        return self.network.parameters()
 
     def __str__(self) -> str:
         """ Allow a nice understanding of what components where used """
