+++ autoPyTorch/pipeline/components/setup/optimizer/RMSpropOptimizer.py
@@ -56,7 +56,7 @@ class RMSpropOptimizer(BaseOptimizerComponent):
 
         # Make sure that input dictionary X has the required
         # information to fit this stage
-        self.check_requirements(X)
+        self.check_requirements(X, y)
 
         self.optimizer = RMSprop(
             params=X['network'].parameters(),
