+++ autoPyTorch/pipeline/components/setup/optimizer/SGDOptimizer.py
@@ -51,7 +51,7 @@ class SGDOptimizer(BaseOptimizerComponent):
 
         # Make sure that input dictionary X has the required
         # information to fit this stage
-        self.check_requirements(X)
+        self.check_requirements(X, y)
 
         self.optimizer = SGD(
             params=X['network'].parameters(),
