+++ autoPyTorch/pipeline/tabular_classification.py
@@ -8,9 +8,9 @@ from sklearn.base import ClassifierMixin
 
 from autoPyTorch.pipeline.base_pipeline import BasePipeline
 from autoPyTorch.pipeline.components.base_choice import autoPyTorchChoice
-from autoPyTorch.pipeline.components.setup.lr_scheduler import SchedulerChoice
-from autoPyTorch.pipeline.components.setup.network import NetworkChoice
-from autoPyTorch.pipeline.components.setup.optimizer import OptimizerChoice
+from autoPyTorch.pipeline.components.setup.lr_scheduler.base_scheduler_choice import SchedulerChoice
+from autoPyTorch.pipeline.components.setup.network.base_network_choice import NetworkChoice
+from autoPyTorch.pipeline.components.setup.optimizer.base_optimizer_choice import OptimizerChoice
 
 
 class TabularClassificationPipeline(ClassifierMixin, BasePipeline):
